feat: add description to traveller category
