+++ src/stacks-hierarchy/Root_PurchaseOverviewScreen/components/Travellers/SingleTravellerSelection.tsx
@@ -1,11 +1,6 @@
 import React from 'react';
 import {UserCountState} from './use-user-count-state';
-import {
-  useTranslation,
-  TicketTravellerTexts,
-  PurchaseOverviewTexts,
-  getTextForLanguage,
-} from '@atb/translations';
+import {useTranslation, PurchaseOverviewTexts} from '@atb/translations';
 import {getReferenceDataName} from '@atb/configuration';
 import {
   RadioGroupSection,
@@ -18,6 +13,7 @@ import {StyleSheet} from '@atb/theme';
 import {HoldingHands} from '@atb/assets/svg/color/images';
 import {useOnBehalfOf} from '@atb/on-behalf-of';
 import {TravelerOnBehalfOfProps} from './types';
+import {getTravellerInfoByFareProductType} from './../utils';
 
 export function SingleTravellerSelection({
   userProfilesWithCount,
@@ -30,7 +26,6 @@ export function SingleTravellerSelection({
   const {t, language} = useTranslation();
   const styles = useStyles();
   const selectedProfile = userProfilesWithCount.find((u) => u.count);
-
   const isOnBehalfOfEnabled = useOnBehalfOf();
 
   const select = (u: UserProfileWithCount) => {
@@ -40,26 +35,6 @@ export function SingleTravellerSelection({
     addCount(u.userTypeString);
   };
 
-  function travellerInfoByFareProductType(
-    fareProductType: string | undefined,
-    u: UserProfileWithCount,
-  ) {
-    const genericUserProfileDescription = getTextForLanguage(
-      u.alternativeDescriptions,
-      language,
-    );
-
-    return [
-      t(
-        TicketTravellerTexts.userProfileDescriptionOverride(
-          u.userTypeString,
-          fareProductType,
-        ),
-      ) || genericUserProfileDescription,
-      t(TicketTravellerTexts.information(u.userTypeString, fareProductType)),
-    ].join(' ');
-  }
-
   return (
     <View>
       <RadioGroupSection<UserProfileWithCount>
@@ -67,7 +42,7 @@ export function SingleTravellerSelection({
         keyExtractor={(u) => u.userTypeString}
         itemToText={(u) => getReferenceDataName(u, language)}
         itemToSubtext={(u) =>
-          travellerInfoByFareProductType(fareProductType, u)
+          getTravellerInfoByFareProductType(fareProductType, u, language, t)
         }
         selected={selectedProfile}
         onSelect={select}
